Move form and list to index route
